requirements.txt
cv2-python
ultralytics
supervision
streamlit
st_pages